Add support for values which are not JSON stringable
Values like undefined and functions are simply dropped by JSON.
This can lead to hashes which match erroneously. The default toString
for these values works fine, so they will be converted to a string
and then bracketed with double curly braces. Thus `undefined` will
not match `"undefined"` and also will not be dropped.