Hash objects with circular references
The replacer function is not a particularly good place to detect and
replace circular references. There is no good way to distinguish between
repeated references that are siblings (and thus stringifiable) and true
circular references where a child references a parent.

So we will instead screen objects before JSON stringifying them. If they
have circular references, they will be replaced with double-curly
strings, and then stringified.