Support node hashing and rebrand to easy-hash
Although Node hashes are already easy for strings, a lot of the work in
this library has gone into stringifying objects for hashing, not just
using the crypto libraries. That work would also be useful in a Node
context.

Removing the browser specific buffer conversion functions as a part of
this shift in focus.