Replace existing security warning with callout
Some commenters have correctly pointed out that security is difficult to
get right and that the consequences for getting it wrong are often
disastrous. One example is this GitHub Issue:

https://github.com/delventhalz/browser-hash/issues/1

While browser-hash was never intended for security purposes and was
always direct about this, the possibility for misunderstanding is real
and must be mitigated as much as possible. To that end, I have reworded
the existing security warning to be more dire, and formatted it as a
callout to more readily catch the eye.

Please do not use browser-hash to hash your users' passwords or as a
substitute for proper signing, encryption, etc.